refactor(realtime): Duplicate old code to legacy 🛠
The goal is to continue to serve legacy code via
subscribe method while implementing the new API.
